+++ src/ui/components/chapterlabel/ChapterLabel.tsx
@@ -1,5 +1,5 @@
 import type { StyleProp, TextStyle } from 'react-native';
-import { Text } from 'react-native'
+import { Text } from 'react-native';
 import React from 'react';
 import { PlayerContext, UiContext } from '../util/PlayerContext';
 import { useChaptersTrack } from '../../hooks/useChaptersTrack';
@@ -13,8 +13,7 @@ export interface ChapterLabelProps {
   /**
    * The playhead position to which the user might seek. Use this property to pass slider values before the actual (debounced) seek happens.
    */
-  scrubTime?: number
-
+  scrubTime?: number;
 }
 
 export interface ChapterLabelState {
@@ -35,17 +34,16 @@ export const DEFAULT_CHAPTER_LABEL_STYLE: TextStyle = {
  * The default time label component for the `react-native-theoplayer` UI.
  */
 
-
 export const ChapterLabel = (props: ChapterLabelProps) => {
-    const currentTime = useCurrentTime();
-    const chapters = useChaptersTrack();
-    const { style, scrubTime } = props;
-    const expectedSeekTarget = scrubTime || currentTime
-    const currentChapter = chapters?.cues?.find(cue => (cue.startTime <= expectedSeekTarget && cue.endTime > expectedSeekTarget))
-    const label = currentChapter?.content
-    return (
-        <PlayerContext.Consumer>
-            {(context: UiContext) => <Text style={[context.style.text, { color: context.style.colors.text }, style]}>{label}</Text>}
-        </PlayerContext.Consumer>
-        );
-}
+  const currentTime = useCurrentTime();
+  const chapters = useChaptersTrack();
+  const { style, scrubTime } = props;
+  const expectedSeekTarget = scrubTime || currentTime;
+  const currentChapter = chapters?.cues?.find((cue) => cue.startTime <= expectedSeekTarget && cue.endTime > expectedSeekTarget);
+  const label = currentChapter?.content;
+  return (
+    <PlayerContext.Consumer>
+      {(context: UiContext) => <Text style={[context.style.text, { color: context.style.colors.text }, style]}>{label}</Text>}
+    </PlayerContext.Consumer>
+  );
+};