+++ src/ui/components/seekbar/thumbnail/SingleThumbnailView.tsx
@@ -3,14 +3,31 @@ import { PlayerContext } from '../../util/PlayerContext';
 import { ThumbnailView } from './ThumbnailView';
 import React, { useContext, useMemo } from 'react';
 import { useThumbnailTrack } from '../../../hooks/useThumbnailTrack';
+import { useSeekable } from '../../../hooks/useSeekable';
+import { useDuration } from '../../../hooks/useDuration';
 
 export interface ThumbnailViewProps {
-  seekableStart: number;
-  seekableEnd: number;
+  /**
+   * The current scrubber time.
+   */
   currentTime: number;
+
+  /**
+   * The width of the seekbar.
+   */
   seekBarWidth: number;
+
+  /**
+   * The size of the thumbnail image, expressed as a percentage of the smaller screen dimension (width or height).
+   *
+   * @default 0.35 (35%).
+   */
+  thumbSize?: number;
 }
 
+/**
+ * Component for displaying a single preview thumbnail image, sourced from a metadata track containing thumbnail data.
+ */
 export function SingleThumbnailView(props: ThumbnailViewProps) {
   const player = useContext(PlayerContext).player;
   const thumbnailTrack = useThumbnailTrack();
@@ -18,17 +35,24 @@ export function SingleThumbnailView(props: ThumbnailViewProps) {
     const window = Dimensions.get('window');
     return 0.35 * Math.min(window.height, window.width);
   }, []);
+  const seekable = useSeekable();
+  const duration = useDuration();
 
   if (!thumbnailTrack) {
     return <></>;
   }
 
-  const { seekableStart, seekableEnd, currentTime, seekBarWidth } = props;
+  const { currentTime, seekBarWidth } = props;
   const marginHorizontal = 8;
+  const normalizedDuration = isNaN(duration) || !isFinite(duration) ? 0 : Math.max(0, duration);
+  const seekableRange = {
+    start: seekable.length > 0 ? seekable[0].start : 0,
+    end: seekable.length > 0 ? seekable[seekable.length - 1].end : normalizedDuration,
+  };
 
   // Do not let the thumbnail pass left & right borders.
-  const seekableRange = seekableEnd - seekableStart;
-  const offset = seekableRange ? (seekBarWidth * (currentTime - seekableStart)) / seekableRange : 0;
+  const range = seekableRange.end - seekableRange.start;
+  const offset = range ? (seekBarWidth * (currentTime - seekableRange.start)) / range : 0;
   let left = -0.5 * thumbnailSize;
   if (offset + marginHorizontal < 0.5 * thumbnailSize) {
     left = -offset - marginHorizontal;