+++ CHANGELOG.md
@@ -10,9 +10,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added an AutoFocusGuide component to be used on tvOS as a wrapper around for example a controlBar. It catches the focus while navigating the UI with the Apple TV remote.
+
 ## [0.12.0] (2025-05-06)
 
 ### Added
+
 - Added localization support.
 - Added support for rendering chapter markers & label.