+++ CHANGELOG.md
@@ -10,6 +10,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added `onScrubbing` callback for the Slider component.
+- Added `scrubTime` property to the `ChapterLabel` component.
+
 ## [0.13.0] (2025-06-11)
 
 ### Added