+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Prevent a fade out of the UI while the skip buttons are being used. 
+- Fixed the useEnded hook to take into account a currentTime that can become slightly larger than the expected stream duration.
 
 ## [0.14.0] (2025-07-04)