+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Prevent a fade out of the UI while the skip buttons are being used.
 - Fixed an issue where the playback rate setting label text didn't fit in its `Text` component.
-- Prevent a fade out of the UI while the skip buttons are being used. 
+- Prevent a fade out of the UI while the skip buttons are being used.
 - Fixed the useEnded hook to take into account a currentTime that can become slightly larger than the expected stream duration.
 - Fixed an issue on Android where a thumbnail preview would contain multiple tiles when the tile image was larger than 2048px.
 - Fixed an issue where `<TimeLabel>` would briefly show an invalid duration when playing a live stream.
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added `onScrubbing` callback for the Slider component.
 - Added `scrubTime` property to the `ChapterLabel` component.
+- Added a `<ConditionalPlayButton>` component that only renders itself while the player is not waiting on media content.
 
 ## [0.13.0] (2025-06-11)